Refactoring timestamp code out of configuration.py and into util.py
